Update README.md
| Field | Type | Label | Description |
| ----- | ---- | ----- | ----------- |
| app_id | [string](#string) |  | fully qualified application ID. In the spirit of github Web URL&lt;-&gt;repo equivalance, the application ID is identified via the URL shown to the user when navigating to the app, namely: https://stacksmith.bitnami.com/foobar/apps/64b0e9e0-c22e-0135-dac4-1a802b3213ed

The actual stacksmith base address can be omitted in case the public SaaS endpoint is used, namely: foobar/apps/64b0e9e0-c22e-0135-dac4-1a802b3213ed

The full URL is useful when users want to point to a private stacksmith instance and allows us to avoid teaching the users to use a flag for that purpose. But the main reason for supporting the full URL to identify the application is that it makes it dead simple for people to copy paste. |
| app_name | [string](#string) |  | If app_id doesn&#39;t exist, use this name when creating it. |
| app_version | [string](#string) |  | if empty, reuses the one of the last revision. |
| files | [Stackerfile.Files](#stacksmith.cli.Stackerfile.Files) |  |  |
| platforms | [Platform](#stacksmith.cli.Platform) | repeated | platforms for which this this application is designed to be built. |
| publishing_regions | [Stackerfile.Region](#stacksmith.cli.Stackerfile.Region) | repeated |  |
| build_options | [Stackerfile.BuildOptions](#stacksmith.cli.Stackerfile.BuildOptions) |  |  |
| base_images | [builder.BaseImages](#builder.BaseImages) |  |  |
| images | [ImageSpec](#stacksmith.cli.ImageSpec) | repeated |  |